Added MPAndroidChart

+++ app/src/main/java/com/example/android/weatherme/data/database/entities/perhour/PerHourEntity.kt
@@ -13,5 +13,5 @@ data class PerHourEntity(
         val lat: Double,
         val lon: Double,
         val timezone: String,
-        val timezoneOffset: Int
+        val timezoneOffset: Long
 ) : Parcelable
